Moving glue to provider.amazon.aws
